requirements.txt
discord.py[voice]
aiohttp>=1.0.0,<1.1.0
websockets>=3.1,<4.0
